chit-chat/requirements.txt
python-socketio==4.4.0
uvicorn==0.11.3